Moved testMVELAccumulate
